This file was conflicted before I committed it. Since it is a binary file the diff was not straight forward. I examined the binary diff using vimdiff, and examined the diffs of the actual function block and header block independently. Only code was added so the updated version is correct.
This is still part of the below commit:
It's been a while since a commit so it's time that I did. Many of the changes are just debug messages that I used while debugging scenarios that caused AODV to act strangely (or so I thought). Also, I have begun to implement the statistics collection routines for the precision of location data.